[ FIX ] : 스플래시 아이콘 사이즈 조정
