+++ app/src/main/java/com/yapp/gallery/SplashActivity.kt
@@ -1,0 +1,31 @@
+package com.yapp.gallery
+
+import android.content.Intent
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+
+import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
+import com.google.firebase.auth.FirebaseAuth
+import com.yapp.gallery.home.screen.home.HomeActivity
+import com.yapp.gallery.login.screen.LoginActivity
+import dagger.hilt.android.AndroidEntryPoint
+import javax.inject.Inject
+
+@AndroidEntryPoint
+class SplashActivity : ComponentActivity() {
+    @Inject lateinit var auth: FirebaseAuth
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        navigate()
+        installSplashScreen().setKeepOnScreenCondition{ false }
+    }
+
+    private fun navigate(){
+        finishAfterTransition()
+        auth.currentUser?.let {
+            startActivity(Intent(this, HomeActivity::class.java))
+        } ?: run{
+            startActivity(Intent(this, LoginActivity::class.java))
+        }
+    }
+}
