--- app/src/main/java/com/yapp/gallery/screen/SplashEffect.kt
@@ -1,6 +1,0 @@
-package com.yapp.gallery.screen
-
-sealed class SplashEffect{
-    object MoveToHome : SplashEffect()
-    object MoveToLogin : SplashEffect()
-}
