+++ app/src/main/res/mipmap-anydpi-v26/ic_splash_icon.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<adaptive-icon xmlns:android="http://schemas.android.com/apk/res/android">
+    <background android:drawable="@color/ic_splash_icon_background"/>
+    <foreground android:drawable="@drawable/ic_splash"/>
+</adaptive-icon>
