--- app/src/main/res/values-v31/themes.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-
-    <style name="Theme.CustomSplashScreenTheme" parent="Theme.YAPPALLRounderTeam2.NoAnim">
-        <item name="android:windowIsTranslucent">true</item>
-        <item name="android:windowSplashScreenBackground">@color/background</item>
-        <item name="android:windowSplashScreenAnimatedIcon">@android:color/transparent</item>
-        <item name="android:windowSplashScreenAnimationDuration">0</item>
-    </style>
-</resources>
