+++ app/src/main/res/values/ic_splash_icon_background.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <color name="ic_splash_icon_background">#161616</color>
+</resources>
