+++ app/src/main/res/values/themes.xml
@@ -16,7 +16,10 @@
     <style name="Theme.YAPPALLRounderTeam2.NoAnim" parent ="Theme.YAPPALLRounderTeam2">
         <item name="android:windowAnimationStyle">@null</item>
     </style>
-    <style name="Theme.CustomSplashScreenTheme" parent="Theme.YAPPALLRounderTeam2.NoAnim">
-        <item name="android:windowIsTranslucent">true</item>
+    <style name="Theme.CustomSplashScreenTheme" parent="Theme.SplashScreen">
+        <item name="android:statusBarColor">@color/background</item>
+        <item name="windowSplashScreenBackground">@color/background</item>
+        <item name="windowSplashScreenAnimatedIcon">@drawable/ic_splash</item>
+        <item name="postSplashScreenTheme">@style/Theme.YAPPALLRounderTeam2.NoAnim</item>
     </style>
 </resources>
